refactor(analytics): use shared database package
