+++ agent/analytics/directory.py
@@ -5,9 +5,9 @@ from uuid import UUID, uuid4
 from sqlalchemy import text
 
 from agent.analytics.capture import fail_soft
-from agent.analytics.database import configured, transaction, workspace_id
 from agent.analytics.identity import opaque_id, opaque_person
 from agent.config import ENV
+from agent.database.analytics import configured, transaction, workspace_id
 
 
 @fail_soft
