+++ agent/analytics/ingestion.py
@@ -7,9 +7,9 @@ from uuid import UUID
 from sqlalchemy import BigInteger, bindparam, text
 from sqlalchemy.ext.asyncio import AsyncConnection
 
-from agent.analytics.database import record_capture, transaction
 from agent.analytics.events import EventEnvelope, EventName
 from agent.config import ENV
+from agent.database.analytics import record_capture, transaction
 
 _INSERT_ID = text(
     "INSERT INTO event_ids(event_id, occurred_at) VALUES (:event_id, :occurred_at) "
