+++ agent/analytics/outbox.py
@@ -8,10 +8,10 @@ from uuid import UUID
 
 from sqlalchemy import text
 
-from agent.analytics.database import record_capture, transaction
 from agent.analytics.events import EventEnvelope
 from agent.analytics.ingestion import ingest
 from agent.config import ENV
+from agent.database.analytics import record_capture, transaction
 
 logger = logging.getLogger(__name__)
